Extract id-message-to-string conversion
